+++ Frontend/userInterface/screens/AssignedRides.tsx
@@ -1,11 +1,15 @@
 import React, {useCallback, useEffect, useRef, useState} from 'react';
-import {Alert, Modal, Platform, StyleSheet, TextInput, TouchableOpacity, View,} from 'react-native';
+import {Alert, Modal, Platform, TextInput, TouchableOpacity, View,} from 'react-native';
 import ThemedText from '@/components/ThemedText';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import {useNavigation} from '@react-navigation/native';
 import {StackNavigationProp} from '@react-navigation/stack';
 import {RootStackParamList} from '@/components/navigation/NavigationTypes';
 import {Ionicons} from "@expo/vector-icons";
+import baseStyles from '../styles/General';
+import rideStyles from '../styles/AssignedRide';
+
+const styles = { ...baseStyles, ...rideStyles };
 
 type AssignedRidesNavigationProp = StackNavigationProp<RootStackParamList, 'AssignedRides'>;
 
@@ -223,54 +227,5 @@ const AssignedRides: React.FC = () => {
         </View>
     );
 };
-
-const styles = StyleSheet.create({
-    container: {flex: 1, padding: 20, backgroundColor: 'white'},
-    emptyContainer: {flex: 1, justifyContent: 'center', alignItems: 'center'},
-    buttonText: {color: 'white', fontSize: 16},
-    endButton: {backgroundColor: 'red', padding: 15, alignItems: 'center', borderRadius: 10, marginTop: 20},
-    messageButton: {backgroundColor: 'blue', padding: 15, alignItems: 'center', borderRadius: 10, marginTop: 10},
-    errorText: {color: 'red', textAlign: 'center', marginTop: 10},
-    labelText: {fontSize: 16, fontWeight: 'bold', color: '#444', marginTop: 10},
-    infoText: {fontSize: 16, color: 'black', marginBottom: 10},
-    header: {flexDirection: 'row', alignItems: 'center', marginBottom: 20,},
-    headerText: {flex: 1, fontSize: 28, textAlign: 'center',},
-    modalContainer: {
-        flex: 1,
-        justifyContent: 'center',
-        alignItems: 'center',
-        backgroundColor: 'rgba(0,0,0,0.5)',
-    },
-    modalContent: {
-        width: '80%',
-        backgroundColor: 'white',
-        padding: 20,
-        borderRadius: 10,
-    },
-    modalTitle: {fontSize: 18, fontWeight: 'bold', marginBottom: 10},
-    inputBox: {
-        height: 50,
-        width: 50,
-        borderColor: '#ddd',
-        borderWidth: 1,
-        borderRadius: 8,
-        fontSize: 24,
-        textAlign: 'center',
-    },
-    modalActions: {flexDirection: 'row', justifyContent: 'space-between'},
-    modalButton: {
-        backgroundColor: '#4a90e2',
-        padding: 10,
-        borderRadius: 10,
-        flex: 1,
-        marginHorizontal: 5,
-        alignItems: 'center'
-    },
-    inputContainer: {
-        flexDirection: 'row',
-        justifyContent: 'space-between',
-        marginBottom: 20,
-    },
-});
 
 export default AssignedRides;
