+++ Frontend/userInterface/screens/CreateAccount.tsx
@@ -1,9 +1,14 @@
 import React, {useState} from 'react';
-import {StyleSheet, Text, TextInput, TouchableOpacity, View} from 'react-native';
+import {Text, TextInput, TouchableOpacity, View} from 'react-native';
 import {useNavigation} from '@react-navigation/native';
 import {StackNavigationProp} from "@react-navigation/stack";
 import {RootStackParamList} from "@/components/navigation/NavigationTypes";
 import ThemedText from "@/components/ThemedText";
+import createBaseStyles from '../styles/CreateAccountCommon';
+import loginStyles from '../styles/GeneralCreate';
+import baseStyles from '../styles/General';
+
+const styles = { ...baseStyles, ...loginStyles, ...createBaseStyles };
 
 type LoginScreenNavigationProp = StackNavigationProp<RootStackParamList, 'Login'>;
 
@@ -148,49 +153,5 @@ const CreateAccount: React.FC = () => {
         </View>
     );
 };
-
-const styles = StyleSheet.create({
-    container: {flex: 1, justifyContent: 'center', padding: 20},
-    input: {
-        height: 40,
-        borderColor: 'black',
-        borderWidth: 1,
-        marginBottom: 15,
-        paddingHorizontal: 10,
-        borderRadius: 10
-    },
-    errorText: {color: 'red', marginBottom: 10},
-    createAccountButton: {backgroundColor: 'blue', padding: 15, alignItems: 'center', borderRadius: 10},
-    createAccountText: {color: 'white'},
-    loginButton: {marginTop: 10, alignItems: 'center'},
-    loginText: {color: 'blue'},
-    label: {fontSize: 16, marginBottom: 5, color: 'black'},
-    header: {flexDirection: 'row', alignItems: 'center', marginBottom: 20,},
-    headerText: {flex: 1, fontSize: 28, textAlign: 'center',},
-    emailContainer: {
-        flexDirection: 'row',
-        alignItems: 'center',
-        borderColor: 'black',
-        borderWidth: 1,
-        borderRadius: 10,
-        padding: 0,
-        marginBottom: 5,
-        height: 35,
-    },
-    emailInput: {
-        flex: 1,
-        color: 'black',
-        fontSize: 16,
-        borderRadius: 10,
-        borderBottomRightRadius: 0,
-        borderTopRightRadius: 0,
-        paddingLeft: 10,
-        height: '100%',
-        width: '90%',
-    },
-    emailDomain: {fontSize: 16, color: 'black', marginRight: 3,},
-    helperText: {fontSize: 12, color: '#b0b0b0', marginBottom: 15,},
-    verticalLine: {width: 2, height: '100%', backgroundColor: 'gray', marginHorizontal: 1,},
-});
 
 export default CreateAccount;
