+++ Frontend/userInterface/screens/DisplayRideInfo.tsx
@@ -1,5 +1,5 @@
 import React, {useEffect, useState} from 'react';
-import {Alert, Modal, Platform, StyleSheet, TextInput, TouchableOpacity, View} from 'react-native';
+import {Alert, Modal, Platform, TextInput, TouchableOpacity, View} from 'react-native';
 import ThemedText from '../components/ThemedText';
 import {StackNavigationProp} from "@react-navigation/stack";
 import {RootStackParamList} from "@/components/navigation/NavigationTypes";
@@ -7,6 +7,10 @@ import {RouteProp, useNavigation} from "@react-navigation/native";
 import CustomDropdown from "@/components/CustomDropdown";
 import {Ionicons} from "@expo/vector-icons";
 import AsyncStorage from "@react-native-async-storage/async-storage";
+import baseStyles from '../styles/General';
+import rideStyles from '../styles/RideInfo';
+
+const styles = { ...baseStyles, ...rideStyles };
 
 
 type DisplayRideInfoNavigationProp = StackNavigationProp<RootStackParamList, 'DisplayRideInfo'>;
@@ -109,7 +113,7 @@ const DisplayRideInfo: React.FC<DisplayRideInfoProps> = ({route}) => {
     };
     const handleSelectReason = (value: React.SetStateAction<string>) => {
         setCancelReason(value);
-        if (value === 'other') {
+        if (value === 'Other') {
             setCancelReason('');
             setShowTextInput(true);
         } else {
@@ -197,37 +201,5 @@ const DisplayRideInfo: React.FC<DisplayRideInfoProps> = ({route}) => {
         </View>
     );
 };
-
-const styles = StyleSheet.create({
-    container: {flex: 1, padding: 20},
-    labelText: {fontSize: 16, fontWeight: 'bold', color: '#444', marginTop: 10},
-    infoText: {fontSize: 16, color: 'black', marginBottom: 10},
-    cancelButton: {backgroundColor: 'red', padding: 15, alignItems: 'center', borderRadius: 10, marginTop: 20},
-    messageButton: {backgroundColor: 'blue', padding: 15, alignItems: 'center', borderRadius: 10, marginTop: 10},
-    buttonText: {color: 'white', fontSize: 16},
-    modalContainer: {flex: 1, justifyContent: 'center', alignItems: 'center', backgroundColor: 'rgba(0,0,0,0.5)'},
-    modalContent: {width: '80%', backgroundColor: 'white', padding: 20, borderRadius: 10},
-    modalTitle: {fontSize: 18, fontWeight: 'bold', marginBottom: 10},
-    inputBox: {
-        height: 100,
-        borderColor: '#ddd',
-        borderWidth: 1,
-        borderRadius: 12,
-        padding: 15,
-        marginBottom: 20,
-        marginTop: 10
-    },
-    modalActions: {flexDirection: 'row', justifyContent: 'space-between', marginTop: 10},
-    modalButton: {
-        backgroundColor: '#4a90e2',
-        padding: 10,
-        borderRadius: 10,
-        flex: 1,
-        marginHorizontal: 5,
-        alignItems: 'center'
-    },
-    header: {flexDirection: 'row', alignItems: 'center', marginBottom: 20,},
-    headerText: {flex: 1, fontSize: 28, textAlign: 'center',},
-});
 
 export default DisplayRideInfo;
