+++ Frontend/userInterface/screens/DriverDashboard.tsx
@@ -1,5 +1,5 @@
 import React, {useEffect, useState} from 'react';
-import {Alert, StyleSheet, TouchableOpacity, View} from 'react-native';
+import {Alert, TouchableOpacity, View} from 'react-native';
 import ThemedText from '@/components/ThemedText';
 import ThemedView from '@/components/ThemedView';
 import Card from '@/components/Card';
@@ -7,6 +7,7 @@ import AsyncStorage from '@react-native-async-storage/async-storage';
 import {useNavigation} from '@react-navigation/native';
 import {StackNavigationProp} from '@react-navigation/stack';
 import {RootStackParamList} from '@/components/navigation/NavigationTypes';
+import styles from '../styles/Dashboard';
 
 type DriverDashboardNavigationProp = StackNavigationProp<RootStackParamList, 'Login'>;
 
@@ -101,22 +102,5 @@ const DriverDashboard: React.FC = () => {
         </ThemedView>
     );
 };
-
-const styles = StyleSheet.create({
-    container: {flex: 1, padding: 20, backgroundColor: 'white'},
-    header: {flexDirection: 'row', alignItems: 'center', marginBottom: 20},
-    titleContainer: {flex: 1, alignItems: 'center'},
-    title: {fontSize: 28, fontWeight: 'bold'},
-    subtitle: {fontSize: 20, marginTop: 5, color: 'gray'},
-    cardsContainer: {flex: 1, justifyContent: 'space-around'},
-    logoutButton: {
-        backgroundColor: 'red',
-        padding: 15,
-        borderRadius: 10,
-        marginTop: 20,
-        alignItems: 'center',
-    },
-    logoutText: {color: 'white', fontSize: 16},
-});
 
 export default DriverDashboard;
