+++ Frontend/userInterface/screens/RiderDashboard.tsx
@@ -1,5 +1,5 @@
-import React, {useCallback, useEffect, useState} from 'react';
-import {StyleSheet, TouchableOpacity, View} from 'react-native';
+import React, {useCallback, useState} from 'react';
+import {TouchableOpacity, View} from 'react-native';
 import ThemedText from '../components/ThemedText';
 import ThemedView from '../components/ThemedView';
 import Card from '../components/Card';
@@ -8,6 +8,7 @@ import {useFocusEffect, useNavigation} from '@react-navigation/native';
 import {StackNavigationProp} from '@react-navigation/stack';
 import {RootStackParamList} from '@/components/navigation/NavigationTypes';
 import {Ionicons} from "@expo/vector-icons";
+import styles from '../styles/Dashboard';
 
 type RiderDashboardNavigationProp = StackNavigationProp<RootStackParamList, 'Login'>;
 
@@ -117,16 +118,5 @@ const RiderDashboard: React.FC = () => {
         </ThemedView>
     );
 };
-
-const styles = StyleSheet.create({
-    container: {flex: 1, padding: 20, backgroundColor: 'white'},
-    header: {flexDirection: 'row', alignItems: 'center', marginBottom: 20},
-    titleContainer: {flex: 1, alignItems: 'center'},
-    title: {fontSize: 28, fontWeight: 'bold'},
-    subtitle: {fontSize: 20, marginTop: 5, color: 'gray'},
-    cardsContainer: {flex: 1, justifyContent: 'space-around'},
-    logoutButton: {backgroundColor: 'red', padding: 15, borderRadius: 10, marginTop: 20, alignItems: 'center'},
-    logoutText: {color: 'white', fontSize: 16},
-});
 
 export default RiderDashboard;
