+++ Frontend/userInterface/screens/SupervisorCreate.tsx
@@ -1,10 +1,15 @@
 import React, {useState} from 'react';
-import {Alert, StyleSheet, Text, TextInput, TouchableOpacity, View} from 'react-native';
+import {Alert, Text, TextInput, TouchableOpacity, View} from 'react-native';
 import {useNavigation} from '@react-navigation/native';
 import {StackNavigationProp} from "@react-navigation/stack";
 import {RootStackParamList} from "@/components/navigation/NavigationTypes";
 import ThemedText from "@/components/ThemedText";
 import {Ionicons} from "@expo/vector-icons";
+import createBaseStyles from '../styles/CreateAccountCommon';
+import superStyles from '../styles/SuperCreate';
+import baseStyles from '../styles/General';
+
+const styles = { ...baseStyles, ...superStyles, ...createBaseStyles };
 
 type LoginScreenNavigationProp = StackNavigationProp<RootStackParamList, 'Login'>;
 
@@ -152,60 +157,5 @@ const SupervisorCreate: React.FC = () => {
         </View>
     );
 };
-
-const styles = StyleSheet.create({
-    container: {flex: 1, justifyContent: 'center', padding: 20},
-    input: {
-        height: 40,
-        borderColor: 'black',
-        borderWidth: 1,
-        marginBottom: 15,
-        paddingHorizontal: 10,
-        borderRadius: 10
-    },
-    errorText: {color: 'red', marginBottom: 10},
-    createAccountButton: {backgroundColor: 'blue', padding: 15, alignItems: 'center', borderRadius: 10},
-    createAccountText: {color: 'white'},
-    radioContainer: {flexDirection: 'row', alignItems: 'flex-start', marginVertical: 15, justifyContent: 'center'},
-    radioButton: {flexDirection: 'row', alignItems: 'center', marginRight: 15},
-    radioOuterCircle: {
-        width: 20,
-        height: 20,
-        borderRadius: 10,
-        borderWidth: 1,
-        borderColor: 'gray',
-        justifyContent: 'center',
-        alignItems: 'center',
-        margin: 10
-    },
-    radioInnerCircle: {width: 10, height: 10, borderRadius: 5, backgroundColor: 'blue'},
-    label: {fontSize: 16, marginBottom: 5, color: 'black'},
-    header: {flexDirection: 'row', alignItems: 'center', marginBottom: 20,},
-    headerText: {flex: 1, fontSize: 28, textAlign: 'center',},
-    emailContainer: {
-        flexDirection: 'row',
-        alignItems: 'center',
-        borderColor: 'black',
-        borderWidth: 1,
-        borderRadius: 10,
-        padding: 0,
-        marginBottom: 5,
-        height: 35,
-    },
-    emailInput: {
-        flex: 1,
-        color: 'black',
-        fontSize: 16,
-        borderRadius: 10,
-        borderBottomRightRadius: 0,
-        borderTopRightRadius: 0,
-        paddingLeft: 10,
-        height: '100%',
-        width: '90%',
-    },
-    emailDomain: {fontSize: 16, color: 'black', marginRight: 3,},
-    helperText: {fontSize: 12, color: '#b0b0b0', marginBottom: 15,},
-    verticalLine: {width: 2, height: '100%', backgroundColor: 'gray', marginHorizontal: 1,},
-});
 
 export default SupervisorCreate;
