+++ Frontend/userInterface/screens/SupervisorEditUser.tsx
@@ -1,11 +1,15 @@
 import React, {useEffect, useState} from 'react';
-import {ActivityIndicator, Alert, Platform, StyleSheet, Text, TextInput, TouchableOpacity, View} from 'react-native';
+import {ActivityIndicator, Alert, Platform, Text, TextInput, TouchableOpacity, View} from 'react-native';
 import ThemedText from '../components/ThemedText';
 import ThemedView from '../components/ThemedView';
 import {useNavigation, useRoute} from '@react-navigation/native';
 import {StackNavigationProp} from "@react-navigation/stack";
 import {RootStackParamList} from "@/components/navigation/NavigationTypes";
 import {Ionicons} from "@expo/vector-icons";
+import baseStyles from '../styles/General';
+import superEdit from '../styles/SuperEdit';
+
+const styles = { ...baseStyles, ...superEdit};
 
 type EditUserPageNavigationProp = StackNavigationProp<RootStackParamList, 'SupervisorEdit'>;
 type RouteParams = { username: string };
@@ -191,63 +195,5 @@ const SupervisorEditUser: React.FC = () => {
         </ThemedView>
     );
 };
-
-const styles = StyleSheet.create({
-    container: {flex: 1, padding: 20},
-    input: {
-        height: 40,
-        borderColor: 'black',
-        borderWidth: 1,
-        marginBottom: 15,
-        paddingHorizontal: 10,
-        borderRadius: 10
-    },
-    radioContainer: {flexDirection: 'row', alignItems: 'flex-start', marginVertical: 15, justifyContent: "center"},
-    radioButton: {flexDirection: 'row', alignItems: 'center', marginRight: 15},
-    selectedRadio: {width: 20, height: 20, borderRadius: 10, backgroundColor: 'blue', marginRight: 10},
-    unselectedRadio: {width: 20, height: 20, borderRadius: 10, borderWidth: 1, borderColor: 'gray', marginRight: 10},
-    updateButton: {backgroundColor: 'blue', padding: 10, alignItems: 'center', borderRadius: 10, marginBottom: 10},
-    deleteButton: {backgroundColor: 'red', padding: 10, alignItems: 'center', borderRadius: 10, marginBottom: 10},
-    backButton: {backgroundColor: 'gray', padding: 15, alignItems: 'center', borderRadius: 10},
-    buttonText: {color: 'white', fontSize: 16},
-    header: {flexDirection: 'row', alignItems: 'center', marginBottom: 20,},
-    headerText: {flex: 0.9, fontSize: 28, textAlign: 'center',},
-    label: {fontSize: 16, marginBottom: 7, color: 'black'},
-    radioOuterCircle: {
-        width: 20,
-        height: 20,
-        borderRadius: 10,
-        borderWidth: 1,
-        borderColor: 'gray',
-        justifyContent: 'center',
-        alignItems: 'center',
-        margin: 10
-    },
-    radioInnerCircle: {width: 10, height: 10, borderRadius: 5, backgroundColor: 'blue'},
-    emailContainer: {
-        flexDirection: 'row',
-        alignItems: 'center',
-        borderColor: 'black',
-        borderWidth: 1,
-        borderRadius: 10,
-        padding: 0,
-        marginBottom: 5,
-        height: 35,
-    },
-    emailInput: {
-        flex: 1,
-        color: 'black',
-        fontSize: 16,
-        borderRadius: 10,
-        borderBottomRightRadius: 0,
-        borderTopRightRadius: 0,
-        paddingLeft: 10,
-        height: '100%',
-        width: '90%',
-    },
-    emailDomain: {fontSize: 16, color: 'black', marginRight: 3,},
-    helperText: {fontSize: 12, color: '#b0b0b0', marginBottom: 15,},
-    verticalLine: {width: 2, height: '100%', backgroundColor: 'gray', marginHorizontal: 1,},
-});
 
 export default SupervisorEditUser;
