+++ Frontend/userInterface/styles/ForgetPassword.ts
@@ -1,0 +1,8 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+    resetButton: {backgroundColor: 'red', padding: 15, alignItems: 'center', borderRadius: 10, marginBottom: 10},
+    resetText: {color: 'white'},
+});
+
+export default styles;
