+++ Frontend/userInterface/styles/GeneralCreate.ts
@@ -1,0 +1,7 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+    loginButton: {marginTop: 10, alignItems: 'center'},
+    loginText: {color: 'blue'},
+});
+export default styles;
