+++ Frontend/userInterface/styles/Login.ts
@@ -1,0 +1,15 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+    logoContainer: {alignItems: 'center', marginBottom: 30},
+    logo: {width: 190, height: 190},
+    inputContainer: {marginBottom: 30},
+    loginButton: {backgroundColor: 'blue', padding: 15, borderRadius: 10, alignItems: 'center', marginBottom: 15,},
+    loginText: {color: 'white', fontSize: 16},
+    createAccountButton: {padding: 15, alignItems: 'center',},
+    createAccountText: {color: 'blue', fontSize: 16},
+    resetButton: {padding: 15, alignItems: 'center'},
+    resetText: {color: 'blue', fontSize: 16,},
+});
+
+export default styles;
