+++ Frontend/userInterface/styles/Privacy.ts
@@ -1,0 +1,9 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+    section: {marginBottom: 20},
+    sectionHeader: {fontSize: 20, fontWeight: 'bold', marginBottom: 10},
+    sectionText: {fontSize: 16, color: '#333'},
+});
+
+export default styles;
