+++ Frontend/userInterface/styles/Settings.ts
@@ -1,0 +1,12 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+    userInfoContainer: {backgroundColor: '#f0f0f0', padding: 10, borderRadius: 5, marginBottom: 20,},
+    userInfoRow: {flexDirection: 'row', alignItems: 'center', marginBottom: 5,},
+    userInfoText: {fontSize: 16, color: '#333', marginLeft: 10,},
+    button: {flexDirection: 'row', alignItems: 'center', backgroundColor: '#4CAF50', padding: 15, marginBottom: 10, borderRadius: 10,},
+    icon: {marginRight: 10},
+    buttonText: {color: 'white', fontSize: 16},
+});
+
+export default styles;
