+++ Frontend/userInterface/styles/ViewReports.ts
@@ -1,0 +1,9 @@
+import {StyleSheet} from "react-native";
+
+const styles = StyleSheet.create({
+    scrollContainer: {paddingBottom: 20},
+    loadingIndicator: {marginTop: 20},
+    squareButton: {width: 35, height: 35, backgroundColor: 'green', alignItems: 'center', justifyContent: 'center', borderRadius: 5,},
+});
+
+export default styles;
